requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
addict==2.4.0
aiohttp @ file:///tmp/build/80754af9/aiohttp_1646806366512/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
astropy==5.2.2
astroquery==0.4.6
astunparse==1.6.3
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
beautifulsoup4==4.12.2
blinker==1.6.2
blobfile==1.3.1
bokeh==3.1.1
brisque==0.0.15
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2022.6.15
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer==2.1.0
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.0.6
cryptography @ file:///tmp/build/80754af9/cryptography_1652083738073/work
cycler==0.11.0
einops==0.4.1
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.3.3
flatbuffers @ file:///home/ktietz/cip/python-flatbuffers_1634039120618/work
fonttools==4.38.0
frozenlist @ file:///tmp/build/80754af9/frozenlist_1637767111923/work
fsspec==2023.9.2
future==0.18.2
gast @ file:///tmp/build/80754af9/gast_1637837472536/work
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///tmp/build/80754af9/grpcio_1637590823556/work
h5py @ file:///tmp/abs_4aewd3wzey/croots/recipe/h5py_1659091371897/work
html5lib==1.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.22.4
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
keras @ file:///home/builder/tkoch/workspace/tensorflow/keras_1659008978158/work/keras-2.8.0-py2.py3-none-any.whl
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
keyring==24.2.0
kiwisolver==1.4.4
libsvm==3.23.0.4
lightning-utilities==0.9.0
lmdb==1.3.0
Markdown @ file:///tmp/build/80754af9/markdown_1614363528767/work
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.2
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
ml-collections==0.1.1
mmcv==2.0.1
mmengine==0.8.5
more-itertools==10.1.0
mpi4py==3.0.3
multidict @ file:///opt/conda/conda-bld/multidict_1640703752579/work
networkx==2.8.8
nibabel==5.1.0
nose==1.3.7
numpy==1.23.2
nvidia-ml-py3==7.352.0
oauthlib @ file:///tmp/abs_08ngfezid4/croots/recipe/oauthlib_1659642459222/work
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
protobuf @ file:///media/imedHDD/imed/personal/xqh/protobuf-3.19.6-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.15.0
pyerfa==2.0.0.3
Pygments==2.16.1
# Editable install with no version control (pyiqa==0.1.5)
-e /media/imedHDD/imed/anaconda3/envs/gxy/lib/python3.8/site-packages
PyJWT @ file:///opt/conda/conda-bld/pyjwt_1657544592787/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytorch-lightning==2.1.0
pytorch-wavelets @ file:///media/imedHDD/imed/personal/gxy/Wave_Net/pytorch_wavelets
pytz==2022.4
pyvo==1.4.2
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.1
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.0
resnest==0.0.5
rfconv==0.0.2b20210406
rich==13.6.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.2
SecretStorage==3.3.3
Shapely==1.8.5.post1
SimpleITK @ file:///media/imedHDD/imed/personal/xqh/SimpleITK-2.2.1-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.4.1
SPLAT==0.32
svm==0.1.0
tb-nightly==2.12.0a20221113
tensorboard @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard_1658918251247/work/tensorboard-2.8.0-py3-none-any.whl
tensorboard-data-server @ file:///tmp/build/80754af9/tensorboard-data-server_1633035064162/work/tensorboard_data_server-0.6.0-py3-none-manylinux2010_x86_64.whl
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow @ file:///home/builder/tkoch/workspace/tensorflow/tensorflow-base_1658921484246/work/tensorflow_pkg/tensorflow-2.8.2-cp38-cp38-linux_x86_64.whl
tensorflow-estimator @ file:///home/builder/tkoch/workspace/tensorflow/tensorflow-estimator_1658918714910/work/tensorflow_estimator-2.8.0-py2.py3-none-any.whl
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.7
torch==1.12.1
torchaudio==0.12.1
torchfile==0.1.0
torchmetrics==1.2.0
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3 @ file:///tmp/abs_5dhwnz6atv/croots/recipe/urllib3_1659110457909/work
visdom==0.1.8.9
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.3.7
wrapt @ file:///tmp/abs_c335821b-6e43-4504-9816-b1a52d3d3e1eel6uae8l/croots/recipe/wrapt_1657814400492/work
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
xyzservices==2023.7.0
yapf==0.32.0
yarl @ file:///tmp/build/80754af9/yarl_1606939922162/work
zipp @ file:///opt/conda/conda-bld/zipp_1652341764480/work
